refactor(decompose): split packages/libraries/graph-model/src/pure/graph/positioning/layout/applyPositions.test.ts into folder
