+++ packages/libraries/graph-model/src/pure/graph/positioning/layout/applyPositions.test/complexCases.test.ts
@@ -1,0 +1,426 @@
+import { describe, it, expect } from 'vitest'
+import * as O from 'fp-ts/lib/Option.js'
+import { applyPositions } from '../applyPositions'
+import { componentsOverlap, type ComponentSubgraph } from '../packComponents'
+import type { Graph, GraphNode, NodeIdAndFilePath, Position } from '../../..'
+import { createGraph, createEmptyGraph } from '../../../construction/createGraph'
+
+describe('applyPositions', () => {
+  describe('random n-ary tree with 100 nodes', () => {
+    it('should position all nodes without edge overlaps', () => {
+      // Use seeded random for deterministic test results
+      const originalRandom: () => number = Math.random
+      Math.random = seededRandom(42)
+
+      // Generate random n-ary tree with 100 nodes
+      const graph: Graph = generateRandomNAryTree(100, 6)
+
+      // Restore original random
+      Math.random = originalRandom
+
+      // Verify initial state: all positions are None
+      Object.values(graph.nodes).forEach(node => {
+        expect(O.isNone(node.nodeUIMetadata.position)).toBe(true)
+      })
+
+      // Apply positions
+      const result: Graph = applyPositions(graph)
+
+      // Assertion 1: All nodes should now have positions
+      const nodesWithPositions: readonly GraphNode[] = Object.values(result.nodes).filter(node =>
+        O.isSome(node.nodeUIMetadata.position)
+      )
+
+      expect(nodesWithPositions.length).toBe(Object.keys(result.nodes).length)
+
+      // Assertion 1.5: No node should be within 10px of any other node
+      // Note: With high branching factors (up to 6 children), angular positioning
+      // may place some nodes close together. This is a known limitation.
+      const positions: readonly Position[] = nodesWithPositions.map(node =>
+        O.toUndefined(node.nodeUIMetadata.position)!
+      )
+      const tooCloseNodes: readonly (readonly [Position, Position])[] = findNodesTooClose(positions, 10)
+      if (tooCloseNodes.length > 0) {
+        console.log(`Found ${tooCloseNodes.length} pairs of nodes too close to each other`)
+        console.log('First 3 examples:', tooCloseNodes.slice(0, 3).map(([p1, p2]) => ({
+          pos1: p1,
+          pos2: p2,
+          distance: Math.sqrt(Math.pow(p2.x - p1.x, 2) + Math.pow(p2.y - p1.y, 2))
+        })))
+      }
+      // Threshold of 35 chosen based on empirical testing with 100-node trees and branching factor of 6
+      // This allows the algorithm some tolerance while still catching regressions
+      expect(tooCloseNodes.length).toBeLessThan(35)
+
+      // Assertion 2: Edge overlaps should be minimal
+      // Note: With high branching factors (up to 6 children) and angular positioning,
+      // some edge overlaps may occur. This is a known limitation of the current algorithm.
+      const edges: readonly Edge[] = extractEdges(result)
+      const overlaps: readonly (readonly [Edge, Edge])[] = findOverlappingEdges(edges)
+
+      if (overlaps.length > 0) {
+        console.log('Found overlapping edges:', overlaps.length)
+      }
+      // Threshold of 60 chosen based on empirical testing with 100-node trees and branching factor of 6
+      // This allows the algorithm some tolerance while still catching regressions
+      expect(overlaps.length).toBeLessThan(60)
+    })
+
+    it('should produce deterministic output for same input', () => {
+      const graph: Graph = generateRandomNAryTree(50, 4)
+
+      const result1: Graph = applyPositions(graph)
+      const result2: Graph = applyPositions(graph)
+
+      // Should produce identical positions
+      Object.keys(result1.nodes).forEach(nodeId => {
+        const pos1: O.Option<Position> = result1.nodes[nodeId]?.nodeUIMetadata.position
+        const pos2: O.Option<Position> = result2.nodes[nodeId]?.nodeUIMetadata.position
+
+        expect(O.isSome(pos1)).toBe(O.isSome(pos2))
+        if (O.isSome(pos1) && O.isSome(pos2)) {
+          expect(pos1.value).toEqual(pos2.value)
+        }
+      })
+    })
+  })
+
+  describe('disconnected components must not overlap (initial-load tangle bug)', () => {
+    // Bug: on first folder load (no positions.json), applyPositions seeds
+    // positions via angular spawning around a ghost root at the origin. With
+    // multiple disconnected roots each having children, all components cluster
+    // around (0,0) and overlap visually. The renderer trusts these positions
+    // verbatim on initial hydration, producing the "insane crazy state" the
+    // user sees until they press Tidy Layout. The fix: applyPositions must
+    // separate overlapping disconnected components before returning.
+
+    it('two disconnected stars with many children must not overlap', () => {
+      // Two roots each with 6 children — angular seeding from a ghost root at
+      // origin places these stars at adjacent quadrants where their bbox
+      // halos collide.
+      const graph: Graph = multiComponentGraph(2, 6)
+      const positioned: Graph = applyPositions(graph)
+      const subgraphs: readonly ComponentSubgraph[] = buildComponentSubgraphs(positioned, 2)
+      expect(componentsOverlap(subgraphs)).toBe(false)
+    })
+
+    it('eight disconnected stars (real-folder shape) must not overlap', () => {
+      // Reproduces the real "insane state" the user sees: many disconnected
+      // roots, each with several children, all radially seeded around origin
+      // and overlapping until Tidy Layout is pressed.
+      const graph: Graph = multiComponentGraph(8, 6)
+      const positioned: Graph = applyPositions(graph)
+      const subgraphs: readonly ComponentSubgraph[] = buildComponentSubgraphs(positioned, 8)
+      expect(componentsOverlap(subgraphs)).toBe(false)
+    })
+
+    it('preserves pre-existing positions on already-positioned nodes', () => {
+      // Component separation must not move nodes that already have a saved
+      // position (e.g., from positions.json). Only the freshly-seeded nodes
+      // get shifted as part of their component.
+      const graph: Graph = createGraph({
+        'pinned_root.md': {
+          ...createNode('pinned_root.md', ['pinned_child.md']),
+          nodeUIMetadata: {
+            color: O.none,
+            position: O.some({ x: 10000, y: 10000 }),
+            additionalYAMLProps: new Map(),
+            isContextNode: false,
+          },
+        },
+        'pinned_child.md': {
+          ...createNode('pinned_child.md', []),
+          nodeUIMetadata: {
+            color: O.none,
+            position: O.some({ x: 10500, y: 10000 }),
+            additionalYAMLProps: new Map(),
+            isContextNode: false,
+          },
+        },
+        // A second component with no positions — applyPositions seeds it.
+        'fresh_root.md': createNode('fresh_root.md', ['fresh_child.md']),
+        'fresh_child.md': createNode('fresh_child.md', []),
+      })
+
+      const positioned: Graph = applyPositions(graph)
+
+      const pinnedRootPos: Position | undefined = O.toUndefined(
+        positioned.nodes['pinned_root.md']?.nodeUIMetadata.position,
+      )
+      const pinnedChildPos: Position | undefined = O.toUndefined(
+        positioned.nodes['pinned_child.md']?.nodeUIMetadata.position,
+      )
+      expect(pinnedRootPos).toEqual({ x: 10000, y: 10000 })
+      expect(pinnedChildPos).toEqual({ x: 10500, y: 10000 })
+    })
+  })
+})
+
+interface Edge {
+  readonly from: Position
+  readonly to: Position
+  readonly id: string
+}
+
+/**
+ * Seeded random number generator for deterministic tests
+ * Uses a simple LCG (Linear Congruential Generator)
+ */
+function seededRandom(seed: number): () => number {
+  let state: number = seed
+  return () => {
+    state = (state * 1664525 + 1013904223) % 4294967296
+    return state / 4294967296
+  }
+}
+
+/**
+ * Create a graph node with no position
+ */
+function createNode(id: NodeIdAndFilePath, outgoingEdges: readonly NodeIdAndFilePath[]): GraphNode {
+  return {
+    kind: 'leaf',
+    absoluteFilePathIsID: id,
+    outgoingEdges: outgoingEdges.map(targetId => ({ targetId, label: '' })),
+    contentWithoutYamlOrLinks: `# ${id}\n\nContent for ${id}`,
+    nodeUIMetadata: {
+      color: O.none,
+      position: O.none,
+      additionalYAMLProps: new Map(),
+      isContextNode: false
+    }
+  }
+}
+
+/**
+ * Generate a random n-ary tree with the specified number of nodes.
+ * Each node has 0 to maxChildren children.
+ */
+function generateRandomNAryTree(nodeCount: number, maxChildren: number): Graph {
+  if (nodeCount <= 0) {
+    return createEmptyGraph()
+  }
+
+  const nodes: Record<NodeIdAndFilePath, GraphNode> = {}
+  const nodeIds: NodeIdAndFilePath[] = []
+
+  // Create root node
+  const rootId: "node_0.md" = 'node_0.md'
+  nodeIds.push(rootId)
+  nodes[rootId] = createNode(rootId, [])
+
+  // Track which nodes can still have children added
+  const availableParents: string[] = [rootId]
+  let nextNodeIndex: number = 1
+
+  // Build tree breadth-first
+  while (nextNodeIndex < nodeCount && availableParents.length > 0) {
+    const parentId: string = availableParents.shift()!
+    const parent: GraphNode = nodes[parentId]
+
+    // Random number of children (0 to maxChildren)
+    const childCount: number = Math.floor(Math.random() * (maxChildren + 1))
+    const children: NodeIdAndFilePath[] = []
+    for (let i: number = 0; i < childCount && nextNodeIndex < nodeCount; i++) {
+      const childId: string = `node_${nextNodeIndex}.md`
+      nodeIds.push(childId)
+      nodes[childId] = createNode(childId, [])
+      children.push(childId)
+      availableParents.push(childId)
+      nextNodeIndex++
+    }
+
+    // Update parent's outgoing edges
+    if (children.length > 0) {
+      nodes[parentId] = {
+        ...parent,
+        outgoingEdges: children.map(targetId => ({ targetId, label: '' }))
+      }
+    }
+  }
+
+  return createGraph(nodes)
+}
+
+function multiComponentGraph(numComponents: number, childrenPerRoot: number): Graph {
+  const nodes: Record<string, GraphNode> = {}
+  for (let c: number = 0; c < numComponents; c++) {
+    const rootId: string = `root_${c}.md`
+    const childIds: readonly string[] = Array.from(
+      { length: childrenPerRoot },
+      (_: unknown, i: number): string => `c${c}_child_${i}.md`,
+    )
+    nodes[rootId] = createNode(rootId, childIds)
+    for (const childId of childIds) {
+      nodes[childId] = createNode(childId, [])
+    }
+  }
+  return createGraph(nodes)
+}
+
+function buildComponentSubgraphs(graph: Graph, numComponents: number): readonly ComponentSubgraph[] {
+  const memberWidth: number = 250
+  const memberHeight: number = 100
+  const buckets: Record<number, GraphNode[]> = {}
+  for (const [id, node] of Object.entries(graph.nodes)) {
+    const compIdx: number = id.startsWith('root_')
+      ? parseInt(id.slice('root_'.length).replace(/\.md$/, ''), 10)
+      : parseInt(id.slice(1, id.indexOf('_')), 10)
+    if (!Number.isFinite(compIdx)) continue
+    buckets[compIdx] ??= []
+    buckets[compIdx].push(node)
+  }
+  const result: ComponentSubgraph[] = []
+  for (let c: number = 0; c < numComponents; c++) {
+    const members: readonly GraphNode[] = buckets[c] ?? []
+    result.push({
+      nodes: members.map(m => {
+        const pos: Position | undefined = O.toUndefined(m.nodeUIMetadata.position)
+        return { x: pos?.x ?? 0, y: pos?.y ?? 0, width: memberWidth, height: memberHeight }
+      }),
+      edges: [],
+    })
+  }
+  return result
+}
+
+/**
+ * Find all pairs of nodes that are too close to each other (within minDistance)
+ */
+function findNodesTooClose(
+  positions: readonly Position[],
+  minDistance: number
+): readonly (readonly [Position, Position])[] {
+  const tooClose: [Position, Position][] = []
+
+  for (let i: number = 0; i < positions.length; i++) {
+    for (let j: number = i + 1; j < positions.length; j++) {
+      const pos1: Position = positions[i]
+      const pos2: Position = positions[j]
+      const distance: number = Math.sqrt(
+        Math.pow(pos2.x - pos1.x, 2) + Math.pow(pos2.y - pos1.y, 2)
+      )
+
+      if (distance < minDistance) {
+        tooClose.push([pos1, pos2])
+      }
+    }
+  }
+
+  return tooClose
+}
+
+/**
+ * Extract all edges from the graph with their positions
+ */
+function extractEdges(graph: Graph): readonly Edge[] {
+  const edges: Edge[] = []
+
+  Object.values(graph.nodes).forEach(node => {
+    const fromPos: Position | undefined = O.toUndefined(node.nodeUIMetadata.position)
+    if (!fromPos) return
+
+    node.outgoingEdges.forEach(edge => {
+      const childNode: GraphNode = graph.nodes[edge.targetId]
+      const toPos: Position | undefined = O.toUndefined(childNode?.nodeUIMetadata.position)
+      if (!toPos) return
+
+      edges.push({
+        from: fromPos,
+        to: toPos,
+        id: `${node.absoluteFilePathIsID}->${edge.targetId}`
+      })
+    })
+  })
+
+  return edges
+}
+
+/**
+ * Find all pairs of overlapping edges
+ */
+function findOverlappingEdges(edges: readonly Edge[]): readonly (readonly [Edge, Edge])[] {
+  const overlaps: [Edge, Edge][] = []
+
+  for (let i: number = 0; i < edges.length; i++) {
+    for (let j: number = i + 1; j < edges.length; j++) {
+      const edge1: Edge = edges[i]
+      const edge2: Edge = edges[j]
+
+      // Skip if edges share a vertex (they're allowed to touch at endpoints)
+      if (shareVertex(edge1, edge2)) {
+        continue
+      }
+
+      if (segmentsIntersect(edge1.from, edge1.to, edge2.from, edge2.to)) {
+        overlaps.push([edge1, edge2])
+      }
+    }
+  }
+
+  return overlaps
+}
+
+/**
+ * Check if two edges share a vertex
+ */
+function shareVertex(edge1: Edge, edge2: Edge): boolean {
+  return (
+    positionsEqual(edge1.from, edge2.from) ||
+    positionsEqual(edge1.from, edge2.to) ||
+    positionsEqual(edge1.to, edge2.from) ||
+    positionsEqual(edge1.to, edge2.to)
+  )
+}
+
+/**
+ * Check if two positions are equal (with small epsilon for floating point)
+ */
+function positionsEqual(p1: Position, p2: Position): boolean {
+  const epsilon: 0.001 = 0.001 as const
+  return Math.abs(p1.x - p2.x) < epsilon && Math.abs(p1.y - p2.y) < epsilon
+}
+
+/**
+ * Check if two line segments intersect (excluding endpoints)
+ * Uses the cross product method
+ */
+function segmentsIntersect(p1: Position, p2: Position, p3: Position, p4: Position): boolean {
+  const d1: number = direction(p3, p4, p1)
+  const d2: number = direction(p3, p4, p2)
+  const d3: number = direction(p1, p2, p3)
+  const d4: number = direction(p1, p2, p4)
+
+  if (((d1 > 0 && d2 < 0) || (d1 < 0 && d2 > 0)) &&
+      ((d3 > 0 && d4 < 0) || (d3 < 0 && d4 > 0))) {
+    return true
+  }
+
+  // Check for collinear cases (edges on same line)
+  if (d1 === 0 && onSegment(p3, p4, p1)) return true
+  if (d2 === 0 && onSegment(p3, p4, p2)) return true
+  if (d3 === 0 && onSegment(p1, p2, p3)) return true
+  if (d4 === 0 && onSegment(p1, p2, p4)) return true
+
+  return false
+}
+
+/**
+ * Calculate the direction of point p3 relative to line segment p1->p2
+ * Returns: positive if p3 is left of the line, negative if right, 0 if collinear
+ */
+function direction(p1: Position, p2: Position, p3: Position): number {
+  return (p3.y - p1.y) * (p2.x - p1.x) - (p2.y - p1.y) * (p3.x - p1.x)
+}
+
+/**
+ * Check if point p lies on line segment p1->p2 (assuming collinear)
+ */
+function onSegment(p1: Position, p2: Position, p: Position): boolean {
+  return (
+    p.x <= Math.max(p1.x, p2.x) &&
+    p.x >= Math.min(p1.x, p2.x) &&
+    p.y <= Math.max(p1.y, p2.y) &&
+    p.y >= Math.min(p1.y, p2.y)
+  )
+}
